Create Breast Cancer Prediction
Coding an algorithm with a high accuracy rate to predict whether a woman's cancer cells (located in her breast) is benign or malignant. The intention of this algorithm is to improve the life expectancy of women by accurately being able to predict a result, in order to have more time to prepare for treatment.